Deal with the input mlir with sext operation and create castOp operation to aievec dialect

+++ test/unit_tests/aievec_tests/i8xi8_conv2d_1x3_after_polygeist_aie-ml/i8xi8.h
@@ -12,6 +12,6 @@ const int kernel_width = 3;
 const int kernel_height = 1;
 
 void conv2d(int8_t *restrict img_in, size_t m1, int8_t *restrict kernel_coeff,
-            size_t m2, int8_t *restrict img_out, size_t m3);
+            size_t m2, int32_t *restrict img_out, size_t m3);
 
 #endif
